frontend/requirements.txt
streamlit
requests
pillow
latex
